requirements.txt
numpy
faiss-cpu
sentence-transformers
torch
pandas
opencv-python

